chore: Rename package to bucket_schema
