+++ cmd/influx/main.go
@@ -198,7 +198,7 @@ var app = cli.App{
 		newWriteCmd(),
 		newBucketCmd(),
 		newCompletionCmd(),
-		newMeasurementSchemaCmd(),
+		newBucketSchemaCmd(),
 	},
 }
