+++ internal/cmd/bucket_schema/csv.go
@@ -1,4 +1,4 @@
-package measurement_schema
+package bucket_schema
 
 import (
 	"fmt"
