+++ internal/cmd/bucket_schema/nsjson_test.go
@@ -1,4 +1,4 @@
-package measurement_schema
+package bucket_schema
 
 import (
 	"strings"
